+++ Sources/SwiftUIKit/Extensions/EdgeInsets+Edge.swift
@@ -10,16 +10,16 @@ import CoreGraphics
 import SwiftUI
 
 public extension EdgeInsets {
-
+    
     /**
      Get the inset for a certain edge.
      */
     func inset(for edge: Edge) -> CGFloat {
         switch edge {
-            case.top: return top
-            case.bottom: return bottom
-            case.leading: return leading
-            case.trailing: return trailing
+        case.top: return top
+        case.bottom: return bottom
+        case.leading: return leading
+        case.trailing: return trailing
         }
     }
 }